+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # UNRELEASED
 - podman upgraded to 0.7.2
 - Kubernetes: 1.11.1
+- Etcd systemd unit default name set to `etcd` instead of `etcd-#{node['name']}`
 
 # 1.11.4 (13.07.2018)
 - CoreDNS: 1.2.0